refactoring split the allmighty ViewController into independent into separate ones.
