+++ src/main/java/io/zeebe/monitor/rest/ExceptionHandler.java
@@ -18,9 +18,11 @@ public class ExceptionHandler {
   private static final Logger LOG = LoggerFactory.getLogger(ExceptionHandler.class);
 
   private final WhitelabelProperties whitelabelProperties;
+  private final WhitelabelPropertiesMapper whitelabelPropertiesMapper;
 
-  public ExceptionHandler(WhitelabelProperties whitelabelProperties) {
+  public ExceptionHandler(WhitelabelProperties whitelabelProperties, WhitelabelPropertiesMapper whitelabelPropertiesMapper) {
     this.whitelabelProperties = whitelabelProperties;
+    this.whitelabelPropertiesMapper = whitelabelPropertiesMapper;
   }
 
   @org.springframework.web.bind.annotation.ExceptionHandler(value = {ClientException.class})
@@ -40,12 +42,7 @@ public class ExceptionHandler {
     model.addAttribute("message", exc.getMessage());
     model.addAttribute("trace", ExceptionUtils.getStackTrace(exc));
 
-    model.addAttribute("custom-title", whitelabelProperties.getCustomTitle());
-    model.addAttribute("context-path", whitelabelProperties.getBasePath());
-    model.addAttribute("logo-path", whitelabelProperties.getLogoPath());
-    model.addAttribute("custom-css-path", whitelabelProperties.getCustomCssPath());
-    model.addAttribute("custom-js-path", whitelabelProperties.getCustomCssPath());
-
+    whitelabelPropertiesMapper.addPropertiesToModel(model, whitelabelProperties);
     return "error";
   }
